+++ bot/utils/messages.py
@@ -7,12 +7,12 @@
 """
 import json
 from enum import Enum
-from typing import Union
+from typing import Callable, Union
 
 from telegram import Message, ParseMode, Sticker, Update, User
 from telegram.ext import CallbackContext
 
-from bot.utils.errors import NoStickerError
+from bot.utils.errors import InputError, NoStickerError
 from bot.utils.logger import StickfixLogger
 
 module_logger = StickfixLogger(__name__)
@@ -61,7 +61,15 @@ def add_error_msg(user: str) -> str:
 
 
 def raise_no_sticker_error(msg: str, cause: str):
-    error = NoStickerError(err_message=msg, err_cause=cause)
+    raise_error(NoStickerError, msg, cause)
+
+
+def raise_input_error(msg: str, cause: str):
+    raise_error(InputError, msg, cause)
+
+
+def raise_error(constructor: Callable, msg: str, cause: str) -> None:
+    error = constructor(err_message=msg, err_cause=cause)
     module_logger.error(error.message)
     module_logger.error(error.cause)
     raise error
@@ -72,3 +80,4 @@ class Commands(str, Enum):
     ADD = "add"
     HELP = "help"
     DELETE_ME = "deleteMe"
+    SET_MODE = "setMode"